refactor: move avatar overlay stuff to separate components
