+++ src/components/Player/Overlay/AvatarOverlayEditor.svelte
@@ -1,124 +1,10 @@
 <script>
-	import Button from '../Common/Button.svelte';
-	import ContentBox from '../Common/ContentBox.svelte';
+	import Button from '../../Common/Button.svelte';
+	import ContentBox from '../../Common/ContentBox.svelte';
+	import {allOverlays} from './overlay';
 
 	export let editModel;
-
-	const OVERLAY_URL = ''; // TODO: replace it with BL CDN
-
-	const allOverlays = [
-		{
-			title: 'The Sun',
-			items: [
-				{
-					title: 'Tier 1',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheSun_Tier1.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheSun_Tier1_preview.webp',
-				},
-				{
-					title: 'Tier 2',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheSun_Tier2.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheSun_Tier2_preview.webp',
-				},
-				{
-					title: 'Tier 3',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheSun_Tier3.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheSun_Tier3_preview.webp',
-					locked: true,
-				},
-			],
-		},
-
-		{
-			title: 'The Moon',
-			items: [
-				{
-					title: 'Tier 1',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheMoon_Tier1.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheMoon_Tier1_preview.webp',
-				},
-				{
-					title: 'Tier 2',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheMoon_Tier2.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheMoon_Tier2_preview.webp',
-				},
-				{
-					title: 'Tier 3',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheMoon_Tier3.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheMoon_Tier3_preview.webp',
-					locked: true,
-				},
-			],
-		},
-
-		{
-			title: 'The Star',
-			items: [
-				{
-					title: 'Tier 1',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheStar_Tier1.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheStar_Tier1_preview.webp',
-				},
-				{
-					title: 'Tier 2',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheStar_Tier2.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheStar_Tier2_preview.webp',
-				},
-				{
-					title: 'Tier 3',
-					url: OVERLAY_URL + '/assets/profile-overlay/TheStar_Tier3.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/TheStar_Tier3_preview.webp',
-					locked: true,
-				},
-			],
-		},
-
-		{
-			title: 'Sparks',
-			items: [
-				{
-					title: 'Tier 1',
-					url: OVERLAY_URL + '/assets/profile-overlay/Sparks_Tier1.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/Sparks_Tier1_preview.webp',
-				},
-				{
-					title: 'Tier 2',
-					url: OVERLAY_URL + '/assets/profile-overlay/Sparks_Tier2.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/Sparks_Tier2_preview.webp',
-				},
-				{
-					title: 'Tier 3',
-					url: OVERLAY_URL + '/assets/profile-overlay/Sparks_Tier3.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/Sparks_Tier3_preview.webp',
-					locked: true,
-				},
-			],
-		},
-
-		{
-			title: 'Special',
-			items: [
-				{
-					title: 'Tier 1',
-					url: OVERLAY_URL + '/assets/profile-overlay/Special_Tier1.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/Special_Tier1_preview.webp',
-					locked: true,
-				},
-				{
-					title: 'Tier 2',
-					url: OVERLAY_URL + '/assets/profile-overlay/Special_Tier2.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/Special_Tier2_preview.webp',
-					locked: true,
-				},
-				{
-					title: 'Tier 3',
-					url: OVERLAY_URL + '/assets/profile-overlay/Special_Tier3.webp',
-					preview: OVERLAY_URL + '/assets/profile-overlay/Special_Tier3_preview.webp',
-					locked: true,
-				},
-			],
-		},
-	];
+	export let roles;
 
 	let fileinput;
 	const changeAvatar = e => {
